Refactor the code to avoid using database
Preeviously we werte obtaining the list of assets and possible pairs
from a local database. Whilst this may indeed the be the best way to
organise the data, for now the requirement for a populated database has
been removed.

An assets folder has been added to store pickled asset prices although I
have added these to the gitignore as they are only used for local
testing purposes